[*] переименовал имя приложения в print-ip; [*] вынес traits-типы в custom_traits.h; [*] добавил тесты на пустые контейнеры, без if constexpr не собиралось; [*] добавил и закомментировал тесты с проверкой генерации ошибки на стадии компиляции при разных типах в кортеже.
