+++ libs/print_ip/tests/main.cpp
@@ -23,21 +23,28 @@ TEST(PrintIpTest, IntegralValues)
 
 TEST(PrintIpTest, StringValue)
 {
+    check(std::string{""}, "");
     check(std::string{"Hello, World!"}, "Hello, World!");
 }
 
 TEST(PrintIpTest, VectorValue)
 {
+    check(std::vector<short>{}, "");
     check(std::vector<int>{100, 200, 300, 400}, "100.200.300.400");
 }
 
 TEST(PrintIpTest, ListValue)
 {
+    check(std::list<int>{}, "");
     check(std::list<short>{400, 300, 200, 100}, "400.300.200.100");
 }
 
 TEST(PrintIpTest, TupleValue)
 {
+    check(std::make_tuple(), "");
     check(std::make_tuple(123, 456, 789, 0), "123.456.789.0");
+
+    // check(std::make_tuple(123, 456.789), ""); // для разных типов выдаётся ошибка на этапе компиляции
+    // check(std::make_tuple(123.0f, 456.789), ""); // для разных типов выдаётся ошибка на этапе компиляции
 }
