+++ main.cpp
@@ -8,7 +8,7 @@ int main()
 {
     using namespace std;
 
-    cout << "My app version: " << PROJECT_VERSION << endl;
+    cout << "print-ip version: " << PROJECT_VERSION << endl;
 
     print_ip(int8_t{-1}) << std::endl;
     print_ip(int16_t{0}) << std::endl;
@@ -16,7 +16,7 @@ int main()
     print_ip(int64_t{8875824491850138409}) << std::endl;
     print_ip(std::string{"Hello, World!"}) << std::endl;
     print_ip(std::vector<int>{100, 200, 300, 400}) << std::endl;
-    print_ip( std::list<short>{400, 300, 200, 100} ) << std::endl;
+    print_ip(std::list<short>{400, 300, 200, 100}) << std::endl;
     print_ip(std::make_tuple(123, 456, 789, 0)) << std::endl;
 
     return 0;
